backend/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
pillow
python-docx
openpyxl
pandas
google-generativeai>=0.8.3
openai
python-dotenv
pytest
httpx
pdf2docx
opencv-python-headless
docx2pdf
comtypes
rembg
numpy
